Debugged automated motion. Store Move and GuidedMove into directory for better organizational clarity.
